requirements.txt
requests
beautifulsoup4
playwright
python-dotenv
python-telegram-bot
